% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCohorts.R
\name{createCohortsCDM}
\alias{createCohortsCDM}
\title{Instantiate cohorts against an omop cdm instance}
\usage{
createCohortsCDM(
  cdm,
  cohortTable,
  type = c("cohorts"),
  cdmDatabaseSchema,
  cohortDatabaseSchema
)
}
\arguments{
\item{cohortTable}{Name of the table to be created where cohorts will be stored}

\item{type}{Which cohorts to create}

\item{cdmDatabaseSchema}{name of the schema where the cdm is stored}

\item{cohortDatabaseSchema}{name of a schema with write access for the creation of cohort table}

\item{connectionDetails}{connection details generated using DatabaseConnector::createConnectionDetails()}
}
\description{
Instantiate cohorts against an omop cdm instance
}
