% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShinyAlopecia}
\alias{runShinyAlopecia}
\title{`runAlopeciaShiny()` launches an app to visualise TreatmentPatterns results for the alopecia study.}
\usage{
runShinyAlopecia()
}
\description{
`runAlopeciaShiny()` launches an app to visualise TreatmentPatterns results for the alopecia study.
}
