% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStudy.R
\name{runStudy}
\alias{runStudy}
\title{run the main elements of the study}
\usage{
runStudy(
  connectionDetails,
  cohortTable,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  instantiateCohorts = FALSE,
  runDiagnostics = FALSE,
  runPatternAnalysis = FALSE,
  outputFolder,
  databaseId,
  minCellCount
)
}
\arguments{
\item{connectionDetails}{connection details generated using DatabaseConnector::createConnectionDetails()}

\item{cohortTable}{Name of the table to be created where cohorts will be stored}

\item{cdmDatabaseSchema}{name of the schema where the cdm is stored}

\item{cohortDatabaseSchema}{name of a schema with write access for the creation of cohort table}

\item{instantiateCohorts}{choose whether to instantiate the cohorts on your database}

\item{runDiagnostics}{choose whether to run the cohort diagnostics}

\item{runPatternAnalysis}{choose whether to run the treatment patterns analysis}

\item{outputFolder}{The folder where the results should be written}

\item{databaseId}{a short name that can identify the database used}

\item{minCellCount}{the minimum number of patients that can be shared for any single count in the results}
}
\description{
run the main elements of the study
}
