% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTreatmentPatterns.R
\name{runTreatmentPatterns}
\alias{runTreatmentPatterns}
\title{Run the treatment patterns analysis}
\usage{
runTreatmentPatterns(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  cohortsGenerated,
  outputSubDir,
  tablePrefix = NULL,
  logger = NULL,
  cohortIds,
  minCellCount = 5
)
}
\arguments{
\item{connectionDetails}{DatabaseConnector connection details.}

\item{cdmDatabaseSchema}{the schema where the cdm is located.}

\item{cohortDatabaseSchema}{A writable location on the database.}

\item{cohortTable}{Name of the main cohort table in characters.}

\item{cohortsGenerated}{A table contianing the output of the createCohorts function, detailing the cohorts that have been instatiated}

\item{outputSubDir}{The output directory where the results should be saved}

\item{tablePrefix}{table prefix}

\item{logger}{logger object}

\item{cohortIds}{the cohortIds to be used form the cohort table}

\item{minCellCount}{minimum cell count}
}
\value{
the drug utilisation results
}
\description{
Run the treatment patterns analysis
}
