% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyAlopecia.R
\name{runShinyAlopecia}
\alias{runShinyAlopecia}
\title{`runAlopeciaShiny()` launches an app to visualise TreatmentPatterns results for the alopecia study.}
\usage{
runShinyAlopecia(resultsFolder = here::here("results"))
}
\arguments{
\item{resultsFolder}{Define the results folder path in character.}
}
\description{
`runAlopeciaShiny()` launches an app to visualise TreatmentPatterns results for the alopecia study.
}
